Clean up gdb debugger initialization, breakpoints and xtensa logging
Esp-gdbstub has been cleaned up and is no longer initialized in the
application. The initialization code was left inline so that it's
easy enough to uncomment when working on a feature.

In the future we can clean this up by wrapping the Makefile inclusions
and C application code in a 'debug' environment variable.

Xtensa debugging has been disabled. It might be worthwhile to explore
what debug levels 1-3 look like, since they might print worthwhile
information for general debugging. Again, this would be nice to
conditionally set when we're inside a "debug" build.