Add exported STL for enclosure back cap
This exported STL models the back cap of the enclosure and features four
M2 metric bolt mounting points for the circuit board.

This model was designed to fit a 2x1 segment of a snappable 2x2 PCB
prototype board. See the README in this directory for exact dimensions.

This part will likely require a built-in fastener since it's the final
piece that must be inserted and fastened when assembling the gadget.