Update enclosure faceplate and body to reduce bezel to 1 millimeter
This exported STL models the front face and body of the enclosure as a
single printable piece.

The previous enclosure mounted the screen and circuit board parallel to
one another, which meant that the wide width of the circuit board would
restrict the display from having a minimally sized bezel. Because the
board will now be mounted on a rack which floats perpendicular to the
screen, the enclosure can be much shorter and narrower at the expense of
added depth.

This design is missing built-in fasteners for both the backplate and
display. These will come in an updated design once this is printed
and tested for fit.