Update enclosure backplate and add mounted circuit board tray
This exported STL models the back cap of the enclosure and features two
M2 metric bolt mounting points for the circuit board tray.

Because the board will now be mounted on a rack which floats
perpendicular to the screen, the enclosure can be much shorter and
narrower (and thus with reduced bezel) at the expense of added depth.

This model was designed to fit a 2x1 segment of a snappable 2x2 PCB
prototype board. See the README in this directory for exact board
dimensions.

It is unclear whether the tray aligns the micro-controller's micro usb
port with the slot in the case. To be tested during assembly.