Add user-friendly initial install Wifi configuration
These changes add the
[HomeACcessoryKid/esp-wifi-config](https://github.com/HomeACcessoryKid/esp-wifi-config)
library, which allows users to easily configure the device's wifi
connection without hard-coding their credentials and flashing the
microcontroller.

Network credentials are stored, and when they are missing, invalid or
if the network doesn't exist, the device will host a Wifi access point.

Users can connect to this access point, visit the local address and
choose and authenticate with their network.